requirements.txt
google-generativeai>=0.5.0
PyPDF2>=3.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.32.0
